+++ src/orderHistory/index.js
@@ -1,6 +1,6 @@
 import * as CoreUtils from "../core/utils";
 import contracts from "../contracts";
-import * as Utils from "./utils";
+import * as OrderUtils from "./utils/orders";
 
 export const getOrders = async (web3, { loanPartyAddress, start, count }) => {
   const b0xContract = await CoreUtils.getContractInstance(
@@ -13,5 +13,22 @@ export const getOrders = async (web3, { loanPartyAddress, start, count }) => {
     .getOrders(loanPartyAddress, start, count)
     .call();
 
-  return Utils.cleanData(data);
+  return OrderUtils.cleanData(data);
+};
+
+export const getLoanPositions = async (
+  web3,
+  { loanPartyAddress, start, count }
+) => {
+  const b0xContract = await CoreUtils.getContractInstance(
+    web3,
+    contracts.B0x.abi,
+    contracts.B0x.address
+  );
+
+  const data = await b0xContract.methods
+    .getLoanPositions(loanPartyAddress, start, count)
+    .call();
+
+  return OrderUtils.cleanData(data);
 };
